Share vector export formatting through a registry
Constraint: Keep the existing vector export behavior while moving JSON and CSV stream formatters behind a Record<string, ExportFormatter>.
Rejected: Adding plugin export plumbing in this task | the dispatch only requested extraction of the built-in formatter registry.
Confidence: high
Scope-risk: narrow
Directive: Add future export formats by registering formatter entries instead of branching inside the route.
Tested: tsc --noEmit; bun test tests/http/vector/; git diff --check; wc -l src/routes/vector/export.ts src/vector/export-formats.ts
Not-tested: Live export against a production-sized vector collection
